refactor: move JsonOptions to config and isJSONRequest to utils
