--- packages/vite/src/node/plugins/json.ts
@@ -1,19 +1,0 @@
-export interface JsonOptions {
-  /**
-   * Generate a named export for every property of the JSON object
-   * @default true
-   */
-  namedExports?: boolean
-  /**
-   * Generate performant output as JSON.parse("stringified").
-   *
-   * When set to 'auto', the data will be stringified only if the data is bigger than 10kB.
-   * @default 'auto'
-   */
-  stringify?: boolean | 'auto'
-}
-
-const jsonLangs = `\\.(?:json|json5)(?:$|\\?)`
-const jsonLangRE = new RegExp(jsonLangs)
-export const isJSONRequest = (request: string): boolean =>
-  jsonLangRE.test(request)
